Move interpret() into the elaborator
